requirements.txt
pillow==11.1.0
pipenv==2023.12.1
Django==4.2.15
django-bootstrap4==25.1
django-crispy-forms==2.3
django-jazzmin==3.0.1
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.0
gunicorn==20.1.0
aiogram==3.10.0




